bbcbasic: ADVAL(-9) returns the full 64-bit microsecond counter
BBC BASIC integers are 64-bit, so do not clamp the timer to the
ioctl return width: the kernel writes the whole counter back
through the callers buffer (whose low word carried the selector)
and adval() grows a long long return. Never wraps in practice.
Falls back gracefully against a mismatched kernel: the previous
31-bit return-value protocol is still honoured, and an older
kernel that ignores -9 reads as zero.

+++ Applications/bbcbasic/bbeval.c
@@ -87,7 +87,7 @@ void getcsr (int*, int*) ;	// Get text cursor (caret) coords
 int vgetc (int, int) ;		// Get character at specified coords
 int oscall (int) ;		// Call an emulated OS function
 int widths (char *, int) ;	// Get string width in graphics units
-int adval (int) ;		// ADVAL function
+long long adval (int) ;		// ADVAL function (64-bit: microsecond timer)
 void *osopen (int, char *) ;	// Open a file
 unsigned char osbget (void*, int*) ; // Get a byte from a file
 long long getptr (void*) ;	// Get file pointer
